Updated Section on how to run
Originally resources for each component had to be referred through your current working directory. I have fixed this issue for both the API and inventory software. As I don't have a copy of Aronium I cannot fully test if changes to the backend will work as intended so this is postponed.